Revert "Enhancment/config seperation"
