+++ README.md
@@ -1,5 +1,5 @@
 [![Build Status](https://travis-ci.com/Aaron-K-T-Berry/generator-ansible-molecule.svg?branch=master)](https://travis-ci.com/Aaron-K-T-Berry/generator-ansible-molecule)
-[![Coverage Status](https://coveralls.io/repos/github/Aaron-K-T-Berry/generator-ansible-molecule/badge.svg?branch=master&service=github)](https://coveralls.io/github/Aaron-K-T-Berry/generator-ansible-molecule?branch=master)
+[![Coverage Status](https://coveralls.io/repos/github/Aaron-K-T-Berry/generator-ansible-molecule/badge.svg?branch=master)](https://coveralls.io/github/Aaron-K-T-Berry/generator-ansible-molecule?branch=master)
 [![Known Vulnerabilities](https://snyk.io/test/github/Aaron-K-T-Berry/generator-ansible-molecule/badge.svg)](https://snyk.io/test/github/Aaron-K-T-Berry/generator-ansible-molecule)
 [![Greenkeeper badge](https://badges.greenkeeper.io/Aaron-K-T-Berry/generator-ansible-molecule.svg)](https://greenkeeper.io/)
