--- generators/app/config/licenses.js
@@ -1,17 +1,0 @@
-const licenses = [
-	{ name: "Apache 2.0", value: "Apache-2.0" },
-	{ name: "MIT", value: "MIT" },
-	{ name: "Mozilla Public License 2.0", value: "MPL-2.0" },
-	{ name: "BSD 2-Clause (FreeBSD) License", value: "BSD-2-Clause-FreeBSD" },
-	{ name: "BSD 3-Clause (NewBSD) License", value: "BSD-3-Clause" },
-	{ name: "Internet Systems Consortium (ISC) License", value: "ISC" },
-	{ name: "GNU AGPL 3.0", value: "AGPL-3.0" },
-	{ name: "GNU GPL 3.0", value: "GPL-3.0" },
-	{ name: "GNU LGPL 3.0", value: "LGPL-3.0" },
-	{ name: "Unlicense", value: "unlicense" },
-	{ name: "No License (Copyrighted)", value: "UNLICENSED" }
-];
-
-module.exports = {
-	licenses
-};
