+++ generators/app/index.js
@@ -6,16 +6,17 @@ const mkdirp = require("mkdirp");
 const path = require("path");
 const _ = require("lodash");
 const exec = require("child_process").exec;
-const prompts = require("./config/prompts").prompts;
-const helper = require("./src/helper");
+const promptBuilder = require("./src/prompts").promptBuilder;
+const getLicenseValue = require("./src/prompts").getLicenseValue;
 const buildPackageJSON = require("./src/package-builder").buildPackageJSON;
-const args = require("./config/arguments").args;
-const options = require("./config/options").options;
+const optionHelper = require("./src/options");
+const argHelper = require("./src/arguments");
 
 module.exports = class extends Generator {
 	constructor(args, opts) {
 		super(args, opts);
-		helper.generateOptions(this);
+		optionHelper.generateOptions(this);
+		argHelper.generateArgs(this);
 	}
 
 	initializing() {}
@@ -29,14 +30,14 @@ module.exports = class extends Generator {
 			)
 		);
 
-		return this.prompt(prompts(this.options)).then(props => {
+		return this.prompt(promptBuilder(this.options)).then(props => {
 			this.props = props;
 		});
 	}
 
 	configuring() {
 		// save options to props
-		const toProcess = [...options, ...args];
+		const toProcess = [...optionHelper.options, ...argHelper.args];
 		for (let item of toProcess) {
 			if (
 				this.options[item.name] !== undefined &&
@@ -145,7 +146,7 @@ module.exports = class extends Generator {
 					authorName: this.props.gitAuthorName,
 					description: this.props.description,
 					metaCompany: this.props.metaCompany,
-					metaLicence: helper.getLicenseValue(this.props.license)
+					metaLicence: getLicenseValue(this.props.license)
 				}
 			}
 		];
