+++ generators/app/src/package-builder.js
@@ -1,10 +1,25 @@
-const pkgTemplate = require("../config/package-json").pkgTemplate;
+const getLicenseValue = require("./prompts").getLicenseValue;
 
 const buildPackageJSON = props => {
-	const newTemplate = { ...pkgTemplate(props) };
+	let pkgTemplate = {
+		name: props.roleName,
+		version: "1.0.0",
+		main: "index.js",
+		repository: props.gitIncludeRepoUrl ? props.gitRepoUrl : "",
+		description: props.description,
+		author: props.gitAuthorName,
+		license: getLicenseValue(props.license),
+		dependencies: {},
+		scripts: {
+			venv: "virtualenv venv",
+			"venv-activate": "source venv/bin/activate",
+			requirements: "pip install -r requirements.txt"
+		}
+	};
+
 	if (props.includeMolecule) {
-		newTemplate.scripts = {
-			...newTemplate.scripts,
+		pkgTemplate.scripts = {
+			...pkgTemplate.scripts,
 			...{
 				check: "molecule check",
 				converge: "molecule converge",
@@ -25,7 +40,7 @@ const buildPackageJSON = props => {
 		};
 	}
 
-	return newTemplate;
+	return pkgTemplate;
 };
 
 module.exports = {
