+++ generators/app/src/prompts.js
@@ -1,8 +1,29 @@
 const gitConfig = require("git-config");
 var gitConfigs = gitConfig.sync();
-const licenses = require("../config/licenses").licenses;
 
-const prompts = options => {
+const licenses = [
+	{ name: "Apache 2.0", value: "Apache-2.0" },
+	{ name: "MIT", value: "MIT" },
+	{ name: "Mozilla Public License 2.0", value: "MPL-2.0" },
+	{ name: "BSD 2-Clause (FreeBSD) License", value: "BSD-2-Clause-FreeBSD" },
+	{ name: "BSD 3-Clause (NewBSD) License", value: "BSD-3-Clause" },
+	{ name: "Internet Systems Consortium (ISC) License", value: "ISC" },
+	{ name: "GNU AGPL 3.0", value: "AGPL-3.0" },
+	{ name: "GNU GPL 3.0", value: "GPL-3.0" },
+	{ name: "GNU LGPL 3.0", value: "LGPL-3.0" },
+	{ name: "Unlicense", value: "unlicense" },
+	{ name: "No License (Copyrighted)", value: "UNLICENSED" }
+];
+
+const getLicenseValue = name => {
+	for (const item of licenses) {
+		if (item.name === name) {
+			return item.value;
+		}
+	}
+};
+
+const promptBuilder = options => {
 	// BASIC
 	const basicQuestionsPrompts = [
 		{
@@ -68,9 +89,7 @@ const prompts = options => {
 		},
 		{
 			when: response => {
-				return (
-					response.includeMolecule || options["include-molecule"] !== undefined
-				);
+				return response.includeMolecule || options["include-molecule"] !== undefined;
 			},
 			type: "confirm",
 			name: "includeCircleCi",
@@ -109,5 +128,6 @@ const prompts = options => {
 
 module.exports = {
 	licenses,
-	prompts
+	getLicenseValue,
+	promptBuilder
 };
